requirements.txt
alembic==1.7.5
anyio==3.4.0
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
cryptography==36.0.0
cycler==0.11.0
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.70.0
fonttools==4.28.2
greenlet==1.1.2
h11==0.12.0
httptools==0.2.0
idna==3.3
kiwisolver==1.3.2
Mako==1.1.6
MarkupSafe==2.0.1
matplotlib==3.5.0
numpy==1.21.4
packaging==21.3
pandas==1.3.4
passlib==1.7.4
Pillow==8.4.0
psycopg2==2.9.2
psycopg2-binary==2.9.2
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.3
PyYAML==6.0
rsa==4.8
scipy==1.7.3
seaborn==0.11.2
setuptools-scm==6.3.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.27
starlette==0.16.0
tomli==1.2.2
typing_extensions==4.0.0
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
